+++ kivyglops.py
@@ -1130,7 +1130,13 @@ class KivyGlop(PyGlop):  # formerly KivyGlop(Widget, PyGlop)
             # can't do it so don't try
             return None
 
-    def prepare_canvas(self, use_meshes=None, axes_index=-1):
+    def prepare_canvas(self, use_meshes=None, xyz_widget_index=-1):
+        '''
+        Keyword arguments:
+        use_meshes -- Set the list of meshes to override `[self._mesh]`.
+        xyz_widget_index -- The index in use_meshes that is the XYZ
+                            widget that shows this model's orientation.
+        '''
         # props = self.properties
         # hitbox = props['hitbox']
         if self._mesh is None:
@@ -2486,7 +2492,7 @@ class KivyGlopsWindow(ContainerForm):  # formerly a subclass of Widget
             for this_glop in self.scene.glops:
                 if this_glop._axes_mesh is not None:
                     this_glop.prepare_canvas([this_glop._axes_mesh],
-                                             axes_index=0)
+                                             xyz_widget_index=0)
                     context = this_glop.get_context()
                     this_glop.set_uniform("texture0_enable", False)
                 else: